rename: queries 디렉터리를 hooks 디렉터리 내로 이동
