+++ tsconfig.json
@@ -34,7 +34,6 @@
       "@/hooks": ["src/hooks/*"],
       "@/mocks": ["src/mocks/*"],
       "@/pages": ["src/pages/*"],
-      "@/queries": ["src/queries/*"],
       "@/styles": ["src/styles/*"],
       "@/types": ["src/types/*"],
       "@/utils": ["src/utils/*"]
